Properly handle paginated results
To find the folder in which to save new patterns, we retrieve a list of
all files and folders created by this application and search it for the
first folder with the specified name ('Stitchbot patterns' by default).

The API response with all of those files and folders, though, is
paginated, and the script hadn't been accounting for this---so when the
script had saved enough patterns that the existing 'Stitchbot patterns'
folder wasn't listed on the first page of results, it created another
folder of the same name (since it wasn't aware of the first's
existence).